more refactoring. working on CTA coponent and background responsiveness
