+++ app/page.tsx
@@ -2,7 +2,7 @@
 
 import Footer from "@/components/Footer";
 import CaseStudy from "@/app/(landing-page)/CaseStudy";
-import Contact from "@/app/(landing-page)/Contact";
+import Contact from "@/app/(landing-page)/CallToAction";
 import Testimonial from "@/app/(landing-page)/Testimonial";
 import { useNav } from "@/context/NavContext";
 import Hero from "@/app/(landing-page)/Hero";
@@ -19,24 +19,29 @@ export default function Home() {
   const { isActive } = useNav();
   return (
     <>
-      <section
-        className={`${isActive ? "blur" : ""} bg-no-repeat bg-cover bg-center`}
-      >
-        <Hero />
-      </section>
-      <section className={`${styles["section-padding"]}  flex justify-center`}>
-        <OurServices />
-      </section>
-      <section className={styles["casestudy-padding"]}>
-        <CaseStudy />
-      </section>
-      <section className={styles["section-padding"]}>
-        <Testimonial />
-      </section>
-      <section className="relative border-3 border-green-400">
-        <Contact />
-      </section>
-
+      <div className="min-h-screen ">
+        <section
+          className={`${
+            isActive ? "blur" : ""
+          } bg-no-repeat bg-cover bg-center`}
+        >
+          <Hero />
+        </section>
+        <section
+          className={`${styles["section-padding"]}  flex justify-center`}
+        >
+          <OurServices />
+        </section>
+        <section className={styles["casestudy-padding"]}>
+          <CaseStudy />
+        </section>
+        <section className={styles["section-padding"]}>
+          <Testimonial />
+        </section>
+        <section className="relative border-3 border-green-400">
+          <Contact />
+        </section>
+      </div>
       <Footer currentPage="landing" />
     </>
   );
