+++ components/scss/CallToAction.module.scss
@@ -1,14 +1,14 @@
 @use "../../app/queries.scss";
 @import "../../app/queries.scss";
 
-.container {
+.CTAContainer {
   display: flex;
   justify-content: center;
   align-items: center;
   height: fit-content;
   // transform: translateY(100px);
 }
-.background-container {
+.imageContainer {
   width: 100%;
   // -height: 840px;
   // transform: translateY(100px);
@@ -19,7 +19,7 @@
   }
 }
 
-.content-container {
+.contentContainer {
   position: absolute;
   display: flex;
   justify-content: flex-end;
@@ -39,7 +39,7 @@
   }
 }
 
-.text-container {
+.textContainer {
   z-index: 10;
   display: flex;
   flex-direction: column;
@@ -60,6 +60,7 @@
     }
   }
   p {
+    text-wrap: pretty;
     color: var(--Primary-Text, #2c2d31);
     font-weight: 400;
     line-height: normal;
@@ -83,5 +84,17 @@
   img {
     max-width: 233px;
     width: 100%;
+  }
+}
+
+.buttonContainer {
+  display: flex;
+  flex-direction: column;
+  gap: 32px;
+  @include mid-break {
+    flex-direction: row;
+  }
+  @include large-break {
+    flex-direction: row;
   }
 }
