+++ components/scss/ProjectDetail.module.scss
@@ -5,9 +5,6 @@
   padding: 80px 72px;
 }
 
-.imageContainer {
-}
-
 .image {
   width: 100%;
   height: 100%;
@@ -231,35 +228,85 @@
   position: absolute;
   top: 0;
   display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  gap: 56px;
-  padding: 104px 72px;
+  grid-template-columns: repeat(1, 1fr);
+  gap: 32px;
+  padding: 48px 32px;
+  @include mid-break {
+    grid-template-columns: repeat(2, 1fr);
+    gap: 56px;
+    padding: 104px 72px;
+  }
+  @include large-break {
+    grid-template-columns: repeat(2, 1fr);
+    gap: 56px;
+    padding: 104px 72px;
+    justify-content: center;
+    align-items: center;
+  }
 
   .textContainer {
+    text-wrap: pretty;
     background-color: white;
     display: flex;
     flex-direction: column;
     justify-content: center;
     border-radius: 16px;
-    padding: 56px 70px;
+    padding: 32px;
     gap: 32px;
+    @include mid-break {
+      padding: 56px 70px;
+      gap: 32px;
+    }
     h1 {
       color: var(--Primary-Text, #2c2d31);
-      font-size: 40px;
       font-style: normal;
       font-weight: 400;
       line-height: normal;
+      font-size: 32px;
+      @include mid-break {
+        font-size: 40px;
+      }
     }
     p {
       color: var(--Secondary-Text, #5c5d6d);
-      font-size: 24px;
+      font-size: 20px;
       font-style: normal;
       font-weight: 400;
       line-height: normal;
-      letter-spacing: -0.96px;
+      letter-spacing: -0.8px;
+      @include mid-break {
+        font-size: 24px;
+        letter-spacing: -0.96px;
+      }
     }
   }
+  .imageContainerTeam {
+    border-radius: 25px;
+    height: 295px;
+    overflow: hidden;
+    Image {
+      height: 100%;
+      width: 100%;
+      object-fit: cover;
+    }
+  }
+}
 
-  .imageContainer {
+.imageContainerLanding {
+  display: flex;
+  justify-content: center;
+}
+
+.buttonContainer {
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+  @include mid-break {
+    flex-direction: row;
+    gap: 32px;
+  }
+  @include large-break {
+    flex-direction: row;
+    gap: 32px;
   }
 }
